requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
boto3==1.39.4
botocore==1.39.4
bottle==0.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
dynamixel-sdk==3.7.31
executing==2.2.0
filelock==3.18.0
fsspec==2025.5.1
getch==1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ikpy==3.0.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.95.0
openai-whisper==20250625
opencv-contrib-python==4.12.0.88
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
playsound==1.2.2
pooch==1.8.2
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pypot==5.0.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.3.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
unrealspeech==0.1.5
urllib3==2.5.0
wcwidth==0.2.13
wget==3.2
wheel==0.45.1
